+++ src/driver/typed_slot_heuristic.rs
@@ -60,9 +60,9 @@ impl Default for FlowDetection {
 pub(super) struct TypedHeuristicSessionSlot<E, P>
 where
     E: FlowExtractor,
-    P: SessionParser,
-    P::Message: Send + 'static,
-    E::Key: Send + 'static,
+    P: SessionParser + Sync,
+    P::Message: Send + Sync + 'static,
+    E::Key: Send + Sync + 'static,
 {
     extractor: E,
     driver: FlowSessionDriver<E, P, ()>,
@@ -78,7 +78,9 @@ where
 impl<E, P> TypedHeuristicSessionSlot<E, P>
 where
     E: FlowExtractor + Clone,
-    P: SessionParser + Clone,
+    P: SessionParser + Clone + Sync,
+    P::Message: Send + Sync + 'static,
+    E::Key: Send + Sync + 'static,
 {
     pub(super) fn new(
         extractor: E,
@@ -89,8 +91,8 @@ where
         monotonic_timestamps: bool,
     ) -> (Self, SlotHandle<P::Message, E::Key>)
     where
-        E::Key: Send + 'static,
-        P::Message: Send + 'static,
+        E::Key: Send + Sync + 'static,
+        P::Message: Send + Sync + 'static,
     {
         let parser_kind = parser.parser_kind();
         let msg_buf = Arc::new(SegQueue::new());
@@ -120,8 +122,8 @@ where
         msg_buf: Arc<SegQueue<SlotMessage<P::Message, E::Key>>>,
     ) -> Self
     where
-        E::Key: Send + 'static,
-        P::Message: Send + 'static,
+        E::Key: Send + Sync + 'static,
+        P::Message: Send + Sync + 'static,
     {
         let parser_kind = parser.parser_kind();
         Self {
@@ -142,9 +144,9 @@ where
 impl<E, P> ErasedSlot<E::Key> for TypedHeuristicSessionSlot<E, P>
 where
     E: FlowExtractor + Send,
-    E::Key: Hash + Eq + Clone + Send + 'static,
-    P: SessionParser + Send + 'static,
-    P::Message: Send + 'static,
+    E::Key: Hash + Eq + Clone + Send + Sync + 'static,
+    P: SessionParser + Send + Sync + 'static,
+    P::Message: Send + Sync + 'static,
 {
     fn track_into(
         &mut self,
@@ -235,9 +237,9 @@ where
 pub(super) struct TypedHeuristicDatagramSlot<E, D>
 where
     E: FlowExtractor,
-    D: DatagramParser,
-    D::Message: Send + 'static,
-    E::Key: Send + 'static,
+    D: DatagramParser + Sync,
+    D::Message: Send + Sync + 'static,
+    E::Key: Send + Sync + 'static,
 {
     extractor: E,
     driver: FlowDatagramDriver<E, D, ()>,
@@ -253,7 +255,9 @@ where
 impl<E, D> TypedHeuristicDatagramSlot<E, D>
 where
     E: FlowExtractor + Clone,
-    D: DatagramParser + Clone,
+    D: DatagramParser + Clone + Sync,
+    D::Message: Send + Sync + 'static,
+    E::Key: Send + Sync + 'static,
 {
     pub(super) fn new(
         extractor: E,
@@ -264,8 +268,8 @@ where
         monotonic_timestamps: bool,
     ) -> (Self, SlotHandle<D::Message, E::Key>)
     where
-        E::Key: Send + 'static,
-        D::Message: Send + 'static,
+        E::Key: Send + Sync + 'static,
+        D::Message: Send + Sync + 'static,
     {
         let parser_kind = parser.parser_kind();
         let msg_buf = Arc::new(SegQueue::new());
@@ -295,8 +299,8 @@ where
         msg_buf: Arc<SegQueue<SlotMessage<D::Message, E::Key>>>,
     ) -> Self
     where
-        E::Key: Send + 'static,
-        D::Message: Send + 'static,
+        E::Key: Send + Sync + 'static,
+        D::Message: Send + Sync + 'static,
     {
         let parser_kind = parser.parser_kind();
         Self {
@@ -317,9 +321,9 @@ where
 impl<E, D> ErasedSlot<E::Key> for TypedHeuristicDatagramSlot<E, D>
 where
     E: FlowExtractor + Send,
-    E::Key: Hash + Eq + Clone + Send + 'static,
-    D: DatagramParser + Send + 'static,
-    D::Message: Send + 'static,
+    E::Key: Hash + Eq + Clone + Send + Sync + 'static,
+    D: DatagramParser + Send + Sync + 'static,
+    D::Message: Send + Sync + 'static,
 {
     fn track_into(
         &mut self,